+++ trace-processor/applications/sink/trace-ingestor/src/test/java/eu/csgroup/coprs/monitoring/traceingestor/UpdateCacheTests.java
@@ -1,0 +1,2 @@
+package eu.csgroup.coprs.monitoring;public class UpdateCacheTests {
+}